+++ ft_memmove.c
@@ -1,9 +1,20 @@
-#include <stdio.h>
-#include <stdlib.h>
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_memmove.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: aalbeza <marvin@42.fr>                     +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2017/01/29 20:41:21 by aalbeza           #+#    #+#             */
+/*   Updated: 2017/01/29 20:42:21 by aalbeza          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
 
-void	*ft_memmove(void *dest,const void *src, size_t n)
+#include "libft.h"
+
+void	*ft_memmove(void *dest, const void *src, size_t n)
 {
-	int			i;
+	size_t		i;
 	const char	*source;
 	char		*destination;
 
@@ -15,5 +26,5 @@ void	*ft_memmove(void *dest,const void *src, size_t n)
 		destination[i] = source[i];
 		i++;
 	}
-	return((char*)dest);
+	return ((char*)dest);
 }
